risolto server che andava in loop

+++ KNN esteso/knn/src/main/java/knn/knn/mining/IKNNService.java
@@ -10,6 +10,6 @@ import knn.knn.database.InsufficientColumnNumberException;
 
 public interface IKNNService {
 
-    public String getModello(Integer id, Integer formato,String nome) throws JsonProcessingException,TrainingDataException,ClassNotFoundException, IOException,InsufficientColumnNumberException, DatabaseConnectionException;
+    public String getModello(Long id, Integer formato,String nome) throws JsonProcessingException,TrainingDataException,ClassNotFoundException, IOException,InsufficientColumnNumberException, DatabaseConnectionException;
 
 }
